+++ NEWS.md
@@ -1,3 +1,7 @@
-# qcbsRworkshops 0.0.0.9000
+# qcbsRworkshops 0.0.1
 
+* Added `build_workshop()` build workshops with option to download files.
+* Added `render_workshops()` to render the workshop.
+* Added `install_workshop_pkgs()` to install packages dependencies stored
+in `pkgs.yml` files.
 * Added a `NEWS.md` file to track changes to the package.